requirements.txt
bokeh==3.0.3
pandas==1.5.2
Pillow==9.4.0
streamlit==1.17.0
